% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiDiSTATIS_F.b.d_Plot.R
\name{DiDiSTATIS_F.b.d_Plot}
\alias{DiDiSTATIS_F.b.d_Plot}
\title{Plot F.b.d (Barycentric Group Partial Factor Scores)}
\usage{
DiDiSTATIS_F.b.d_Plot(res_DiDiSTATIS, axes = c(1, 2))
}
\arguments{
\item{res_DiDiSTATIS}{The output of DiDiSTATIS}

\item{axes}{Axes to plot}
}
\description{
Plot F.b.d (Barycentric Group Partial Factor Scores)
}
