% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiDiSTATIS_Fab.d_Plot.R
\name{DiDiSTATIS_Fab.d_Plot}
\alias{DiDiSTATIS_Fab.d_Plot}
\title{Plot Fab.d (Discriminant Group Partial Factor Scores)}
\usage{
DiDiSTATIS_Fab.d_Plot(res_DiDiSTATIS, axes = c(1, 2))
}
\arguments{
\item{res_DiDiSTATIS}{The output of DiDiSTATIS}

\item{axes}{Axes to plot}
}
\description{
Plot Fab.d (Discriminant Group Partial Factor Scores)
}
