% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_DiDiSTATIS_Boot_centered_CIs_BD.R
\name{Plot_DiDiSTATIS_Boot_centered_CIs_BD}
\alias{Plot_DiDiSTATIS_Boot_centered_CIs_BD}
\title{Plot histogram of permtuted squared distance from Groups to Grand Compromise}
\usage{
Plot_DiDiSTATIS_Boot_centered_CIs_BD(res_DiDiSTATIS, axes = c(1, 2),
  ellipse = TRUE, percentage = 0.95, Flip_axis1 = FALSE,
  Flip_axis2 = FALSE, dev.new = TRUE)
}
\arguments{
\item{res_DiDiSTATIS}{The output of DiDiSTATIS}

\item{axes}{Axes to plot}

\item{ellipse}{Boolean for ellipse or convex hull}

\item{percentage}{1 - alpha cutoff}

\item{Flip_axis1}{Flag to flip Component 1}

\item{Flip_axis2}{Flag to flip Component 2}

\item{dev.new}{Flag to appease ReporteRs (set FALSE to print results to pptx)}
}
\value{
Histograms of the Dev2's
}
\description{
Plot histogram of permtuted squared distance from Groups to Grand Compromise
}
