% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_DiDiSTATIS_F.b.d.R
\name{Plot_DiDiSTATIS_F.b.d}
\alias{Plot_DiDiSTATIS_F.b.d}
\title{Plot F.b.d (Barycentric Group Partial Factor Scores)}
\usage{
Plot_DiDiSTATIS_F.b.d(res_DiDiSTATIS, axes = c(1, 2))
}
\arguments{
\item{res_DiDiSTATIS}{The output of DiDiSTATIS}

\item{axes}{Axes to plot}
}
\description{
Plot F.b.d (Barycentric Group Partial Factor Scores)
}
