% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_DiDiSTATIS_F.bCd.R
\name{Plot_DiDiSTATIS_F.bCd}
\alias{Plot_DiDiSTATIS_F.bCd}
\title{Plot F.bCd (Discriminant Group Partial Factor Scores)}
\usage{
Plot_DiDiSTATIS_F.bCd(res_DiDiSTATIS, axes = c(1, 2), b = 1, d = 1,
  Flip_axis1 = FALSE, Flip_axis2 = FALSE, dev.new = TRUE)
}
\arguments{
\item{res_DiDiSTATIS}{The output of DiDiSTATIS}

\item{axes}{Axes to plot}

\item{b}{The Category to plot}

\item{d}{The Group to plot}

\item{Flip_axis1}{Flag to flip Component 1}

\item{Flip_axis2}{Flag to flip Component 2}

\item{dev.new}{Flag to appease ReporteRs (set FALSE to print results to pptx)}
}
\description{
Plot F.bCd (Discriminant Group Partial Factor Scores)
}
