% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_DiDiSTATIS_FAB...R
\name{Plot_DiDiSTATIS_FAB..}
\alias{Plot_DiDiSTATIS_FAB..}
\title{Plot FAB.. (Discriminant Grand Factor Scores)}
\usage{
Plot_DiDiSTATIS_FAB..(res_DiDiSTATIS, axes = c(1, 2), Flip_axis1 = FALSE,
  Flip_axis2 = FALSE, dev.new = TRUE)
}
\arguments{
\item{res_DiDiSTATIS}{The output of DiDiSTATIS}

\item{axes}{Axes to plot}

\item{Flip_axis1}{Flag to flip Component 1}

\item{Flip_axis2}{Flag to flip Component 2}

\item{dev.new}{Flag to appease ReporteRs (set FALSE to print results to pptx)}
}
\description{
Plot FAB.. (Discriminant Grand Factor Scores)
}
