% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_DiDiSTATIS_Fab.d.R
\name{Plot_DiDiSTATIS_Fab.d}
\alias{Plot_DiDiSTATIS_Fab.d}
\title{Plot Fab.d (Discriminant Group Partial Factor Scores)}
\usage{
Plot_DiDiSTATIS_Fab.d(res_DiDiSTATIS, axes = c(1, 2), priority = "ab",
  connect = 0)
}
\arguments{
\item{res_DiDiSTATIS}{The output of DiDiSTATIS}

\item{axes}{Axes to plot}

\item{priority}{#"ab", "d"}

\item{connect}{#0 = all, integer or vector connects those ab or d (depending on priority)}
}
\description{
Plot Fab.d (Discriminant Group Partial Factor Scores)
}
