% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_DiDiSTATIS_Perm_r2_Plain_b_.d.R
\name{Plot_DiDiSTATIS_Perm_r2_Plain_b_.d}
\alias{Plot_DiDiSTATIS_Perm_r2_Plain_b_.d}
\title{Plot histogram of permtuted r^2_Plain_b_.d}
\usage{
Plot_DiDiSTATIS_Perm_r2_Plain_b_.d(res_DiDiSTATIS, dev_new = TRUE)
}
\arguments{
\item{res_DiDiSTATIS}{The output of DiDiSTATIS}

\item{dev_new}{Flag to appease ReporteRs (set FALSE to print results to pptx)}
}
\value{
Histogram
}
\description{
Plot histogram of permtuted r^2_Plain_b_.d
}
