% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_DiMDS_Fab.R
\name{Plot_DiMDS_Fab}
\alias{Plot_DiMDS_Fab}
\title{Plot Fab}
\usage{
Plot_DiMDS_Fab(res_DiMDS, axes = c(1, 2))
}
\arguments{
\item{res_DiMDS}{The output of DiMDS}

\item{axes}{Axes to plot, default = c(1,2)}
}
\description{
Plot Fab
}
