% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_DiMDS_Perm_Rows_r2plain.b.R
\name{Plot_DiMDS_Perm_Rows_r2plain.b}
\alias{Plot_DiMDS_Perm_Rows_r2plain.b}
\title{Plot results from test of Null Hypothesis on Rows}
\usage{
Plot_DiMDS_Perm_Rows_r2plain.b(res_DiMDS)
}
\arguments{
\item{res_DiMDS}{Output from DiMDS_Perm_Rows}
}
\value{
Histogram of the r2 under the null
}
\description{
Plot results from test of Null Hypothesis on Rows
}
