% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_DiMDS_Perm_Rows_r2plain.disc_perm.R
\name{Plot_DiMDS_Perm_Rows_r2plain.disc_perm}
\alias{Plot_DiMDS_Perm_Rows_r2plain.disc_perm}
\title{Plot results from test of Null Hypothesis on Rows}
\usage{
Plot_DiMDS_Perm_Rows_r2plain.disc_perm(res_DiMDS, Title = "r2plain.disc_perm")
}
\arguments{
\item{res_DiMDS}{Output from DiMDS_Perm_Rows}

\item{Title}{Name the figure}
}
\value{
Histogram of the r2 under the null
}
\description{
Plot results from test of Null Hypothesis on Rows
}
