% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_DiMDS_SSab.R
\name{Plot_DiMDS_SSab}
\alias{Plot_DiMDS_SSab}
\title{Plot SSab}
\usage{
Plot_DiMDS_SSab(res_DiMDS, axes = c(1, 2), dev_new = F)
}
\arguments{
\item{res_DiMDS}{The output of DiMDS}

\item{axes}{Axes to plot, default = c(1,2)}

\item{dev_new}{Flag to create a new device}
}
\description{
Plot SSab
}
