% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_DiSTATIS_F_Plain.R
\name{Plot_DiSTATIS_F_Plain}
\alias{Plot_DiSTATIS_F_Plain}
\title{Plot DiSTATIS Compromise Factor Scores}
\usage{
Plot_DiSTATIS_F_Plain(res_DiSTATIS, axes = c(1, 2), Flip_axis1 = FALSE,
  Flip_axis2 = FALSE, dev.new = TRUE)
}
\arguments{
\item{res_DiSTATIS}{The output of EigenDiSTATIS}

\item{axes}{Axes to plot, by default c(1,2)}

\item{Flip_axis1}{Flag to flip Component 1}

\item{Flip_axis2}{Flag to flip Component 2}

\item{dev.new}{Flag to appease ReporteRs (set FALSE to print results to pptx)}
}
\description{
Plot DiSTATIS Compromise Factor Scores
}
