% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_HiDiSTATIS_Boot_centered_CIs.R
\name{Plot_HiDiSTATIS_Boot_centered_CIs}
\alias{Plot_HiDiSTATIS_Boot_centered_CIs}
\title{Plot histogram of permtuted squared distance from Groups to Grand Compromise}
\usage{
Plot_HiDiSTATIS_Boot_centered_CIs(res_HiDiSTATIS, axes = c(1, 2), i = 1,
  ellipse = TRUE, percentage = 0.95, dev.new = TRUE)
}
\arguments{
\item{res_HiDiSTATIS}{The output of HiDiSTATIS}

\item{axes}{Axes to plot}

\item{i}{The stimulus to plot}

\item{ellipse}{Boolean for ellipse or convex hull}

\item{dev.new}{Flag to appease ReporteRs (set FALSE to print results to pptx)}
}
\value{
Histograms of the Dev2's
}
\description{
Plot histogram of permtuted squared distance from Groups to Grand Compromise
}
