% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_HiDiSTATIS_Boot_centered_example.R
\name{Plot_HiDiSTATIS_Boot_centered_example}
\alias{Plot_HiDiSTATIS_Boot_centered_example}
\title{Plot histogram of permtuted squared distance from Groups to Grand Compromise}
\usage{
Plot_HiDiSTATIS_Boot_centered_example(res_HiDiSTATIS, i = 1, dev.new = TRUE)
}
\arguments{
\item{res_HiDiSTATIS}{The output of HiDiSTATIS}

\item{i}{The stimulus to plot}

\item{dev.new}{Flag to appease ReporteRs (set FALSE to print results to pptx)}
}
\value{
Histograms of the Dev2's
}
\description{
Plot histogram of permtuted squared distance from Groups to Grand Compromise
}
