% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_HiDiSTATIS_F...R
\name{Plot_HiDiSTATIS_F..}
\alias{Plot_HiDiSTATIS_F..}
\title{Plot F.. (HiDiSTATIS Grand Compromise Factor Scores)}
\usage{
Plot_HiDiSTATIS_F..(res_HiDiSTATIS, axes = c(1, 2), Flip_axis1 = FALSE,
  Flip_axis2 = FALSE, dev.new = TRUE)
}
\arguments{
\item{res_HiDiSTATIS}{The output of HiDiSTATIS}

\item{axes}{Axes to plot}

\item{Flip_axis1}{Flag to flip Component 1}

\item{Flip_axis2}{Flag to flip Component 2}

\item{dev.new}{Flag to appease ReporteRs (set FALSE to print results to pptx)}
}
\value{
Grand Compromise Factor Scores
}
\description{
Plot F.. (HiDiSTATIS Grand Compromise Factor Scores)
}
