% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_HiDiSTATIS_Perm_d2.R
\name{Plot_HiDiSTATIS_Perm_d2}
\alias{Plot_HiDiSTATIS_Perm_d2}
\title{Plot histogram of permtuted squared distance from Groups to Grand Compromise}
\usage{
Plot_HiDiSTATIS_Perm_d2(res_HiDiSTATIS, dev_new = TRUE)
}
\arguments{
\item{res_HiDiSTATIS}{The output of HiDiSTATIS}

\item{dev_new}{Flag to appease ReporteRs (set FALSE to print results to pptx)}
}
\value{
Histograms of the Dev2's
}
\description{
Plot histogram of permtuted squared distance from Groups to Grand Compromise
}
