% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_HiDiSTATIS_Perm_r2_Groups.R
\name{Plot_HiDiSTATIS_Perm_r2_Groups}
\alias{Plot_HiDiSTATIS_Perm_r2_Groups}
\title{Plot histogram of permtuted r^2_Groups}
\usage{
Plot_HiDiSTATIS_Perm_r2_Groups(res_HiDiSTATIS, dev_new = TRUE)
}
\arguments{
\item{res_HiDiSTATIS}{The output of HiDiSTATIS}

\item{dev_new}{Flag to appease ReporteRs (set FALSE to print results to pptx)}
}
\value{
Histograms of the r^2s
}
\description{
Plot histogram of permtuted r^2_Groups
}
