% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_HMFA_F.d.R
\name{Plot_HMFA_F.d}
\alias{Plot_HMFA_F.d}
\title{Plot F.d (Group Factor Scores)}
\usage{
Plot_HMFA_F.d(res_HMFA, axes = c(1, 2), d = 1, Flip_axis1 = FALSE,
  Flip_axis2 = FALSE, dev.new = TRUE)
}
\arguments{
\item{res_HMFA}{The output of HMFA}

\item{axes}{Axes to plot}

\item{d}{Which group to plot}

\item{Flip_axis1}{Flag to flip Component 1}

\item{Flip_axis2}{Flag to flip Component 2}

\item{dev.new}{Flag to appease ReporteRs (set FALSE to print results to pptx)}
}
\description{
Plot F.d (Group Factor Scores)
}
