% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotConfusion_norm_SignedCtrb.R
\name{PlotConfusion_norm_SignedCtrb}
\alias{PlotConfusion_norm_SignedCtrb}
\title{Plot the signed contribution to Chi2}
\usage{
PlotConfusion_norm_SignedCtrb(Confusion_mat_norm = NULL, DESIGN_rows = NULL,
  scale_max_100 = TRUE, dev_new = TRUE)
}
\arguments{
\item{Confusion_mat_norm}{Confusion matrix, row-normed}

\item{DESIGN_rows}{Design info for the rows}

\item{scale_max_100}{Boolean to set the max to 100 percent}

\item{dev_new}{Flag to appease ReporteRs (set FALSE to print results to pptx)}
}
\value{
A pretty confusion matrix
}
\description{
Plot the signed contribution to Chi2
}
