% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotConfusion_SignedCtrb.R
\name{PlotConfusion_SignedCtrb}
\alias{PlotConfusion_SignedCtrb}
\title{Plot the signed contribution to Chi2}
\usage{
PlotConfusion_SignedCtrb(Confusion_mat = NULL, is.percent = FALSE)
}
\arguments{
\item{Confusion_mat}{Confusion matrix}
}
\value{
A pretty confusion matrix
}
\description{
Plot the signed contribution to Chi2
}
