% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print_res2ppt_DiSTATIS.R
\name{Print_res2ppt_DiSTATIS}
\alias{Print_res2ppt_DiSTATIS}
\title{Print standard DiSTATIS results}
\usage{
Print_res2ppt_DiSTATIS(res_DiSTATIS, axes = c(1, 2), Fik = 1,
  Flip_axis1 = F, Flip_axis2 = F, main = NULL)
}
\arguments{
\item{res_DiSTATIS}{The output of DiDiSTATIS}

\item{axes}{Axes to plot}

\item{Fik}{For Fik, which stimuli to plot}

\item{Flip_axis1}{Flag to flip Component 1}

\item{Flip_axis2}{Flag to flip Component 2}

\item{main}{Title}
}
\value{
A Powerpoint of results
}
\description{
Print standard DiSTATIS results
}
