% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_DiDiSTATIS_F.b...R
\name{Plot_DiDiSTATIS_F.b..}
\alias{Plot_DiDiSTATIS_F.b..}
\title{Plot F.b.. (Barycentric Grand Factor Scores)}
\usage{
Plot_DiDiSTATIS_F.b..(res_DiDiSTATIS, axes = c(1, 2))
}
\arguments{
\item{res_DiDiSTATIS}{The output of DiDiSTATIS}

\item{axes}{Axes to plot}
}
\description{
Plot F.b.. (Barycentric Grand Factor Scores)
}
