% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_DiDiSTATIS_Fab...R
\name{Plot_DiDiSTATIS_Fab..}
\alias{Plot_DiDiSTATIS_Fab..}
\title{Plot Fab.. (Discriminant Grand Factor Scores)}
\usage{
Plot_DiDiSTATIS_Fab..(res_DiDiSTATIS, axes = c(1, 2))
}
\arguments{
\item{res_DiDiSTATIS}{The output of DiDiSTATIS}

\item{axes}{Axes to plot}
}
\description{
Plot Fab.. (Discriminant Grand Factor Scores)
}
