% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Hist.R
\name{Plot_Hist}
\alias{Plot_Hist}
\title{Plot histogram}
\usage{
Plot_Hist(Distribution, alpha = 0.05, Score, Score_colors = "blue",
  breaks = 40, pad = 1/5, main = NULL, dev_new = TRUE)
}
\arguments{
\item{Distribution}{The distribution of values}

\item{alpha}{The alpha threshold}

\item{Score}{The score(s) to compare to the distribution}

\item{Score_colors}{The score(s) to compare to the distribution}

\item{pad}{A fraction (of the range), to pad xlim}

\item{main}{Title}

\item{dev_new}{Flag to appease ReporteRs (set FALSE to print results to pptx)}
}
\value{
Histogram
}
\description{
Plot histogram
}
